fix(mpi): quita versión anterior del controller
